Add /forceOut

+++ app/services/remove_participant_service.rb
@@ -3,19 +3,34 @@
 class RemoveParticipantService < ApplicationService
   # @attr_reader params [Hash]
   # - message: [Telegram::Bot::Types::Message] Incoming message
+  # - participant: [Participant] Participant to cancel translation
 
   include DocumentsApiConcern
   include SemanticsConcern
+  include ParticipantConcern
 
   def call
     return if !message || !chat || !document || !document.pending?
 
-    result = []
     if doc_participant.persisted?
+      caption = 'Вы больше не участвуете'
+      if param_participant
+        return "У вас нет прав на принудительное исключение переводчиков" unless is_admin?(participant)
+
+        caption = "#{doc_participant.participant.full_name} больше не участвует"
+        notifications = []
+        doc_participant.participant.subscriptions.each do |subscription|
+          notifications << {
+            text: "Вас исключили из числа переводчиков документа #{document.url}",
+            chat_id: subscription.chat_id
+          }
+          NotificationsService.perform(notifications: notifications)
+        end
+      end
       doc_participant.destroy
-      "Вы больше не участвуете в переводе этого документа. Сейчас добровольцев: #{load_participants_count}"
+      "#{caption} в переводе этого документа. Сейчас добровольцев: #{load_participants_count}"
     else
-      "Вы не участвуете в переводе этого документа. Сейчас добровольцев: #{load_participants_count}"
+      "#{param_participant ? "#{param_participant} не участвует" : 'Вы не участвуете'} в переводе этого документа. Сейчас добровольцев: #{load_participants_count}"
     end
   end
 
@@ -23,6 +38,10 @@ class RemoveParticipantService < ApplicationService
 
   def message
     params[:message]
+  end
+
+  def param_participant
+    params[:participant]
   end
 
   def chat
@@ -41,10 +60,14 @@ class RemoveParticipantService < ApplicationService
     )
   end
 
+  def waiting_participant
+    @waiting_participant ||= param_participant ? load_participant : participant
+  end
+
   def doc_participant
     @doc_participant ||= DocumentParticipant.find_or_initialize_by(
       document_id: document.id,
-      participant_id: participant.id
+      participant_id: waiting_participant&.id
     )
   end
